@@ -12,11 +12,13 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.0.4",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6",
     "react-scripts": "4.0.3",
+    "react-transition-group": "^4.4.2",
+    "tinycolor2": "^1.4.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
