@@ -1,15 +1,20 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import reportWebVitals from './reportWebVitals';
+import { HashRouter } from "react-router-dom";
+import { createRoot } from 'react-dom/client';
+
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 
-ReactDOM.render(
-  <React.StrictMode>
+
+const container = document.getElementById('root');
+const root = createRoot(container); // createRoot(container!) if you use TypeScript
+root.render( 
+  <HashRouter>
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
+  </HashRouter>
 );
+
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
